feat(services): HeadsetService provides methods to get headset info
Note that this has only been tested with an Arctis 7+,  the _knownHeadsets list has been retrieved from other projects but should be working fine.
A future update might allow to add new headsets from a public property, as this service could be use in other projects.

Worth noting this code has been done quickly, is not optimized and might behave unexpectedly in some occasions. Tests will be written in the future.