feat: battery monitoring logic
Logic for the app to work properly. Main logic is running on a separate thread to avoid blocking the app thread while performing operations, this allows notifyIcon to stay responsive when refresh operations are running in the background (i.e always).

New functionalities will be added such as the possibility to deactivate notifications or choose the right headset if several are found.